mm, page_alloc: Initialize subsection maps for non-NUMA systems
Subsection maps are only initialized for NUMA systems when
they invoke free_areas_init(). Not initializing the subsection
maps with memory that exists at boot can cause the memmap for
a section to be freed when memory for a subsection within that
section is hot-added and then hot-removed, which is not correct
behavior, as pfn_valid() will always fail for any memory within
the section, since it no longer has a memmap. Thus, initialize
subsection maps for non-NUMA systems as well.